Export `MessageHistoryMutatedWarning` from `pydantic_ai.exceptions`
