+++ docs/message-history.md
@@ -514,7 +514,7 @@ Adjacent part-style items (user content and [`ModelRequestPart`][pydantic_ai.mes
 Both `enqueue` methods return an `enqueue_id` (`str`) for a non-empty call, or `None` when called with no content. When the queued content is actually delivered into run history, the [event stream](agent.md#streaming-all-events) yields an [`EnqueuedMessagesEvent`][pydantic_ai.messages.EnqueuedMessagesEvent] carrying that `enqueue_id` and the delivered messages (exactly as they landed in history), so a client can observe when its steering message took effect. The event carries the delivered message objects themselves — the same objects held in the run's message history. A history processor that replaces history with new message objects does not affect the event, but in-place mutation of a delivered message will be visible through it.
 
 !!! warning "Don't mutate existing messages in place"
-    `enqueue` and [history processors](#processing-message-history) change the conversation by adding or building *new* message objects, and are the supported ways to do so. Mutating a message that's already part of the history in place (e.g. `ctx.messages[0].parts[0].content = '...'` from a tool) is not supported: [instrumentation](logfire.md) caches each message's serialized form for the duration of a run, so the `gen_ai.input.messages` attribute recorded on later model request spans may not reflect the mutation. When this is detected at the end of a run, a `MessageHistoryMutatedWarning` is emitted; the run-level `pydantic_ai.all_messages` attribute always reflects the final history.
+    `enqueue` and [history processors](#processing-message-history) change the conversation by adding or building *new* message objects, and are the supported ways to do so. Mutating a message that's already part of the history in place (e.g. `ctx.messages[0].parts[0].content = '...'` from a tool) is not supported: [instrumentation](logfire.md) caches each message's serialized form for the duration of a run, so the `gen_ai.input.messages` attribute recorded on later model request spans may not reflect the mutation. When this is detected at the end of a run, a [`MessageHistoryMutatedWarning`][pydantic_ai.exceptions.MessageHistoryMutatedWarning] is emitted; the run-level `pydantic_ai.all_messages` attribute always reflects the final history.
 
 ### From inside a tool or hook
