+++ pydantic_ai_slim/pydantic_ai/__init__.py
@@ -32,6 +32,7 @@ from .exceptions import (
     ConcurrencyLimitExceeded,
     FallbackExceptionGroup,
     IncompleteToolCall,
+    MessageHistoryMutatedWarning,
     ModelAPIError,
     ModelHTTPError,
     ModelRetry,
@@ -199,6 +200,7 @@ __all__ = (
     'ModelHTTPError',
     'FallbackExceptionGroup',
     'IncompleteToolCall',
+    'MessageHistoryMutatedWarning',
     'SkipModelRequest',
     'SkipToolExecution',
     'SkipToolValidation',
